[encryption] Add encryption support to Env
This commit introduces reading and writing encrypted files through Env.
At this stage, encryption is not usable, as it uses a static dummy key.

Originally, I wanted to use a simple Caesar's cipher like ROT13 until
the rest of the key infra is in place, but then I decided this proof of
concept implementation should also prove that CTR, which is our chosen
mode of operation[1] works for our access patterns, so I ended up using
OpenSSL even in this proof of concept, albeit with a static key and
key size.

[1] https://s.apache.org/kudu-atrest-enc-design

Change-Id: Id5a81b5736a0c1970818dbbc4a234480476b8d41
Reviewed-on: http://gerrit.cloudera.org:8080/17251
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>