[build] drop RHEL/CentOS 6 OpenSSL ABI change workaround
Changelist 10f525519 introduced a workaround to an OpenSSL ABI
compatibility break between RHEL/CentOS 6.4 and later versions.
RHEL/CentOS 6 is beyond its EOL date already and Kudu doesn't
officially support it since 1.14 release [1], so this patch removes
the workaround accordingly.

Nevertheless, even with this patch, Kudu can still be compiled and
run on RHEL/CentOS 6, as before.  The only change is that now it
compiles against the OpenSSL which came with the OS at the build
machine, not the OpenSSL which came with RHEL/CentOS 6.4.

[1] https://kudu.apache.org/releases/1.14.0/docs/release_notes.html

Change-Id: I014d93b3b953d11281b928de9de43eec8ab4e997
Reviewed-on: http://gerrit.cloudera.org:8080/17264
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>