KUDU-3223: Management of per-table level limit
When the table's size or row exceeds a threshold, the insert and update
privilege will be removed from the authz token, only scan and delete
privilege are allowed.

Admin is responsible to change the on_disk_size_limit or
row_count_limiti per-table. -1 means no limit.

This feature depends on authz token expiration, and it caused a delay
for write forbidden. Anyway, it provides an approximately management for
write.

Change-Id: I2dbf365ad59f17c0a4e2e7ea6a5afaa7680724b0
Reviewed-on: http://gerrit.cloudera.org:8080/17273
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>