[hms] allow hive.metastore.disallow.incompatible.col.type.changes to be true
A major distribution of Hive and Kudu (CDP) has changed its default
behavior to set hive.metastore.disallow.incompatible.col.type.changes to
true. This is a safer behavior for Hive, but breaks the Kudu-HMS
integration, since we require the configuration for DROP COLUMN
operations, and check the Hive configuration as such at startup.

HIVE-24987 tracks the work on the Hive-side to permit Kudu operations to
alter tables in a seemingly type-incompatible way that is actually safe
for Kudu.

Until then, this adds an option to disable this Kudu-side check (kept
the current behavior by default, but added a gflag to make it more
easily configurable).

Change-Id: Idc9777e134e243abfe768d631592541720e101a8
Reviewed-on: http://gerrit.cloudera.org:8080/17296
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>