KUDU-2612: send next txn keepalive sooner in case of timeout
This patch updates the code of the C++ client to set the timeout for
TransactionKeepAlive() RPC to be half of the target period for sending
keepalive messages.  Also, if the RPC with the prior keepalive message
timed out, sending next keepalive message sooner gives more chances for
the transaction to survive before it's automatically aborted by the
backend due to not receiving keepalive messages for long time.

This patch also makes the corresponding pieces of the Kudu C++ and
the Java clients consistent (see https://gerrit.cloudera.org/#/c/17305).

Change-Id: Ic12c5615152e61ee81e4f48e4978d4c5b1fa9828
Reviewed-on: http://gerrit.cloudera.org:8080/17310
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>