[thirdparty] enable building curl with alternative OpenSSL
If building curl in thirdparty when an alternative OpenSSL library
is installed and pointed to by PATH and pkgconfig, the following
compilation might happen:

  thirdparty/src/curl-7.68.0/lib/vtls/openssl.c: In function ‘Curl_ossl_seed’:
  thirdparty/src/curl-7.68.0/lib/vtls/openssl.c:500:15: error: implicit declaration of function ‘RAND_egd’; did you mean ‘RAND_add’? [-Werror=implicit-function-declaration]
       int ret = RAND_egd(data->set.str[STRING_SSL_EGDSOCKET]?
		 ^~~~~~~~
		 RAND_add
  cc1: some warnings being treated as errors
  make[1]: *** [Makefile:2067: libcurl_la-openssl.lo] Error 1

The issue was in the configuration script -- it correctly found path
for header files and the shared objects of the custom OpenSSL library,
but since the system OpenSSL library was there as well, the order
of paths in LDFLAGS made the "configure" script picking up the default
library instead of the custom one.  The issue manifests itself only in
environments where the difference between the default/OS OpenSSL library
and the custom one contains some substantial difference in the API or
build/configuration options.  For this particular case, those were
'OpenSSL 1.0.1e-fips 11 Feb 2013' vs 'OpenSSL 1.1.1i  8 Dec 2020'
for the default/OS and the custom OpenSSL library correspondingly.

This patch addresses the issue by patching the configure.in script.

Change-Id: Ia47c60b0dcef0212c777d628ab8c2f0be9c0a368
Reviewed-on: http://gerrit.cloudera.org:8080/17318
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>