KUDU-2612: an extra test for txn keepalive failover in Java client
This is a follow-up to 096f1ddf09047ea11d78a661010dd549ffa9af51.

This patchs adds an extra test scenario similar the one added
in the prior changelist, but with additional twist of "rolling"
unavailability of leader masters.  In addition, it verifies that
RPC error responses from TxnManager due to the unavailability
of TxnStatusManager are properly handled by the Java client.

Change-Id: Ib278d402bee85fb0442cbce98b2b4ab022229eb4
Reviewed-on: http://gerrit.cloudera.org:8080/17321
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins