[client] retry master RPCs on network errors
AsyncLeaderMasterRpc and AsyncRandomTxnManagerRpc fail fast when it runs
into a network error. We expect such errors in a couple of scenarios:
- the client is misconfigured and is pointing at the wrong masters
- the masters are down or otherwise unavailable

The current handling makes it convenient in the former case, since users
can immediately adjust their application or tooling with the correct
masters list. However, this seems to be a far less expected scenario
than the latter. So, this patch updates the behavior to retry on such
errors, as we would an RPC timeout. The end result is that
leader-master-bound requests should be more resilient to master
failures.

The primary motivation for this is to make TxnManager calls more robust
to master failures -- a later patch will leverage this behavior.

Change-Id: Iae2febb9e890acf6f7efd5cce3cb7e4f7b5f683d
Reviewed-on: http://gerrit.cloudera.org:8080/17343
Tested-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>