KUDU-2612: more info on abandoned transaction tracking
This patch cross-references the --txn_staleness_tracker_interval_ms
flag in the description of the --txn_keepalive_interval_ms one.
That's useful for better understanding of how the abandoned transaction
tracking works and how to properly change its parameters or disable it,
if necessary.

This patch doesn't contain any functional changes.

Change-Id: I5b626f1f7e384efd616019013559f171ad7eaddc
Reviewed-on: http://gerrit.cloudera.org:8080/17341
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>