[hms] KUDU-1884 Add support for custom SASL protocol name
This patch adds support to set a custom SASL protocol name
when using the KuduMetastorePlugin in the HMS. Because
there isn’t a good way to pass configuration the plugin
today, it uses an environment variable instead.

The patch updates the master_hms-itest to use the new
custom SASL protocol name feature to validate functionality.

Change-Id: Id8d9e208da1d767bb24470ab031a5266461d070b
Reviewed-on: http://gerrit.cloudera.org:8080/17350
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Grant Henke <granthenke@apache.org>