[doc] KUDU-2181 Update multi-master addition/removal/recovery documentation
This change updates the documentation for:
1) "Migrating to Multiple Masters" that uses the new
`kudu master add` CLI command merged recently.

2) "Removing Kudu Masters from a Multi-Master Deployment"
that uses the `kudu master remove` CLI tool.

3) "Recovering from a dead Kudu Master in a Multi-Master Deployment"
that uses a combination of master remove and add CLI tools.

This change doesn't include any version specific steps
as this doc is meant for the latest Kudu version 1.15.0.
The idea is to introduce an index page for documentation that
points to version specific docs.

This change also removes Cloudera Manager(CM) specific instructions
as they could change with automation in CM.

The rendered version of the doc can be viewed here:
https://github.com/bbhavsar/kudu/blob/bankim/r5/docs/administration.adoc#migrate_to_multi_master

Change-Id: I6a1d5bc6bbf4bc3e82e7046469d2682bf016d3a8
Reviewed-on: http://gerrit.cloudera.org:8080/17352
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins