[util] Add special handling for empty strings in FastHash
Impala uses special handling for empty strings while hashing
https://github.com/apache/impala/blob/master/be/src/runtime/raw-value.inline.h#L352

This leads to discrepancy between hash values for empty
strings between Impala and Kudu.
This change basically matches the behavior with Impala.

Changes to the implementation of hashing function for empty
strings generates a different value and hence incorrect
results across different Kudu client and server versions.
To resolve this issue, a new feature flag BLOOM_FILTER_PREDICATE_V2 has
been added and the older one BLOOM_FILTER_PREDICATE has been marked as
deprecated.

Since FastHash is not used for any persistent data structures
so there are no backward compatibility issues that needs special
handling during upgrade.

Discovered bugs in the Java implementation when a negative
byte is promoted to a long and left shifted. Discovery
was because of the magic value for empty strings/buffer.
Thankfully the Java implementation of the FastHash
is not used currently.

Also fixed a bug in predicate test where flush would be missed
due to increase in number of values which affected TSAN.

Change-Id: Idb2d401ab692decd5bc03b48a96d710546c0039e
Reviewed-on: http://gerrit.cloudera.org:8080/17376
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>