[client] return TimedOut if the deadline has passed in master RPCs
There was a narrow window in which we could return a NetworkError to
users of the master-bound RPCs instead of a TimedOut error. Without this
change I saw ClientTest.NoTxnManager fail 1/300 times and
TxnStatusTableITest.TestSystemClientMasterDown fail ~1/3 of the time
running in a loop via dist-test.

With this change, I ran it with no issue 1000/1000 times each.

This is a follow-up to 96047a8d861d61673b9e2589930eccd78a16e483.

Change-Id: I309ba70d62a48b4dc8fb0e8f30f532593c254860
Reviewed-on: http://gerrit.cloudera.org:8080/17367
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>