KUDU-2612: re-open TxnSystemTable if transaction is in new range
This patch makes the TxnSystemClient on tablet servers retry if attempts
to register the participant yield a NotFound error. This happens when
new ranges are added to the transactions status table.

To disambiguate between the case where the transaction itself doesn't
exist on the TxnStatusManager, this patch also adjust the return value
for such cases to be InvalidArugment, which seems equally fitting for
writing as a part of a bogus transaction ID.

Change-Id: I3af58dcaa3a995dac9dc937c7dfcb652bf004873
Reviewed-on: http://gerrit.cloudera.org:8080/17368
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>