[test] fix flakiness in RegistrationTest.TestTSRegisters
This patch fixes a flakiness in RegistrationTest.TestTSRegisters.

I noticed that the test scenario failed in a pre-commit verification
build for one of my in-flight patches.  I took a quick look and found
that the scenario had a history of flakiness, indeed [1].

Apparently, there was a mistake in the code, but it was concealed by a
race condition.  Once I added a 100ms pause just before the line
this patch removes, the test scenario started failing all the time.

[1] http://dist-test.cloudera.org:8080/test_drilldown?test_name=registration-test

Change-Id: Ic6b072409d2a68bc2ad621c1c021b0f4b3143fb2
Reviewed-on: http://gerrit.cloudera.org:8080/17377
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>