KUDU-2612: update C++ client API to commit a transaction
This patch updates the signature of the KuduTransaction::Commit()
method to address recent feedback on the txn-related API. The idea is
to make the API easier to use, since txn.Commit(false) looks a bit vague
and might be confusing as well.

In essence, the 'wait' parameter is gone and now there are two methods:
  * KuduTransaction::Commit()
  * KuduTransaction::StartCommit()

The former starts committing a multi-row transaction and waits for the
commit phase to finalize.  The latter just starts the commit phase for
a multi-row transaction and returns, not awaiting for the commit phase
to finalize.

Change-Id: Iecac338a753e559597a9348e68c9b09813cc8105
Reviewed-on: http://gerrit.cloudera.org:8080/17392
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>