KUDU-2612: update Java client API to commit a transaction
This patch updates the signature of the KuduTransaction::commit()
method to address recent feedback on the txn-related API.  The idea is
to make the API easier to use, since txn.commit(false) looks a bit vague
and might be confusing as well.

In essence, the 'wait' parameter is gone and now there are two methods:
  * KuduTransaction.commit()
  * KuduTransaction.startCommit()

The former starts committing a multi-row transaction and waits for the
commit phase to finalize.  The latter just starts the commit phase for
a multi-row transaction and returns, not awaiting for the commit phase
to finalize.

These changes mirror recent changes in the C++ client API.

Change-Id: Ia8c48b4d375945649c48428401f09ec5c7cc8cb7
Reviewed-on: http://gerrit.cloudera.org:8080/17393
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>