[util] Add a function to fetch non-default flags as a map
We already have a function that returns non-default flags
as a vector of strings. This change adds a function
that returns non-default flags as a map refactoring
and reusing the existing logic.

This function is used in add master tool to bring
up the new master as it's unnecessary to pass all
the flags to the kudu master. This is similar to
master run implementation in master_runner.cc.

Change-Id: I7e7c9fc4de9be05110dba5ca02f48d46b6b474b7
Reviewed-on: http://gerrit.cloudera.org:8080/17400
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>