KUDU-3164: Add table comment support to Kudu
This patch adds the basic plumbing for table comments,
synchronizing comments with HMS both using the
notification log listener and via tooling, and setting the
comment on CREATE TABLE and ALTER TABLE requests in
the C++ client.

The maximum comment length is 256 characters by default
which aligns with HMS/Apache Impala maximum comment
lengths, but it's configurable with the
max_table_comment_length flag.

Supporting this in the Java and Python clients, and backup
and restore jobs will come in follow on patches.

Change-Id: Ic4dbf4018c38fdd0c7f748f4dfc26816f22bd5b2
Reviewed-on: http://gerrit.cloudera.org:8080/17399
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>