[flags] Remove the costly stringstream to string conversion in a loop
stringstream.str() returns a copy of the underlying string
buffer which is expensive if done repeatedly in a loop.
This change avoids the check for emptiness of the stringstream
in a loop and simply appends the newline character at the
end of every flag. This means there'll be a newline at
the end of the string which is okay.

Change-Id: I6f85aa85feaac34abfa8a582f936392d0a4b0ad6
Reviewed-on: http://gerrit.cloudera.org:8080/17418
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins