[txns][tools] show details of a transaction
This adds a tool to show information related to a single transaction,
including metadata from all of its participants. The goal of this tool
is to be useful in debugging slow commits or inconsistencies with a
transaction.

Currently this outputs info stored in participant's metadata for the
transaction.

$ kudu txn show 0.0.0.0:7051 0
 txn_id | user  |   state   |        commit_datetime
--------+-------+-----------+-------------------------------
 0      | awong | COMMITTED | Fri, 14 May 2021 19:15:18 GMT

            tablet_id             |     begin_commit_datetime     |        commit_datetime
----------------------------------+-------------------------------+-------------------------------
 9204d93c0af14292843b6d432d281fb5 | Fri, 14 May 2021 19:15:18 GMT | Fri, 14 May 2021 19:15:18 GMT

The new tool abides by the --columns flag, as done with the
`kudu txn list` tool, with some additional options for
participant-specific columns.

Change-Id: I9cc5c23b6b46ee75e38aaffe4773881a1ece7294
Reviewed-on: http://gerrit.cloudera.org:8080/17391
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>