[docs][site] Add links to documentation on the releases index page
Kudu documentation is versioned and we have separate docs for every
release. However we don't have an index page that points
to documentation of different releases.

The idea was to introduce a separate docs page that links
docs for every Kudu version. However every Kudu version release page
includes links to the documentation as "Additional Links" and we
already have an index page for Kudu releases.

So instead, it'd be better to add documentation links
to the existing releases index page. In order to limit the size
of releases index page, these documentation links are added only for
currently supported releases.

The documentation links for the archived releases can be fetched
from the release page for that particular archived release like
earlier.

Change-Id: I7dae7a40a73a44649eca9eda2dac646a71e784f5
Reviewed-on: http://gerrit.cloudera.org:8080/17439
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Bankim Bhavsar <bankim@cloudera.com>