[txns] add a means to get transaction metadata from participant
This patch adds a GET_METADATA participant RPC and a method to call it
through the TxnSystemClient. The RPC currently returns the TxnMetadataPB
for the given transaction, giving a bit of insight into the state of the
given participant.

I will use this in an upcoming tool that drills into a transaction and
its participants.

Change-Id: I4b03f13f174bd9a83609fb7ed6106746777b4207
Reviewed-on: http://gerrit.cloudera.org:8080/17446
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-on: http://gerrit.cloudera.org:8080/17467
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>