KUDU-2612 automatically flush sessions on txn commit (Java client)
With this patch, all transactional sessions created off a transction
handle are automatically flushed upon calling commit() on the handle.

As for the KuduTransaction.startCommit() method, it's now necessary
to flush all the transactional sessions created off the transaction
handle before calling the method, otherwise a NonRecoverableException
based on Status.Incomplete() is thrown.

This patch also contains several test scenarios for the newly introduced
functionality.

This changelist is a Java client's counterpart for 45ca93f0e.

Change-Id: I0c52578dd736906cf2610c8cc58496381a9b73ec
Reviewed-on: http://gerrit.cloudera.org:8080/17452
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>