[kserver] output info on server-wide size limit
With this patch, Kudu servers output an INFO log line about the maximum
number of threads in a server-wide thread pool.  That's useful when
assessing how many threads a Kudu server might legitimately spawn.
Yes, those numbers can be calculated given the setting for RLIMIT_NPROC
and values in /proc/sys/kernel/pid_max and /proc/sys/kernel/threads-max,
but (a) those settings might change after the start of a server and
(b) it's convenient to have a pre-computed number ready in the logs.

Change-Id: I62b4168fc153e693f0fcc3d6d1ecfc6be0e69b47
Reviewed-on: http://gerrit.cloudera.org:8080/17474
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>