[java] a property to show output while running a test
While working on non-trivial and relatively long running test scenarios,
I found it's inconvenient that the test output isn't shown while a test
scenario is running.  This patch introduces a new command-line property
'showTestOutput' to direct gradle outputting stdout and stderr from the
JVM on the console, so it's possible to do the standard stream
redirection and other stream manipulation in a command line shell.

An example of running a particular test scenario with the output
redirected to the console:

  ./gradlew -DshowTestOutput cleanTest :kudu-client:test --test ...

By default the property isn't set, so this patch preserves the original
behavior w.r.t. the output of running Java tests.

Change-Id: I82590957004eef427a19f36a32e903012ea220d4
Reviewed-on: http://gerrit.cloudera.org:8080/17479
Tested-by: Kudu Jenkins
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>