[java] utility to detect sanitizer type used for kudu CLI
This patch adds a new test utility to determine the type of the C++
sanitizer type used by the kudu CLI binary, if any.  The new utility is
added into the kudu-test-utils package.  I found it's useful to find
whether the mini-cluster binaries are built with TSAN/ASAN sanitizers,
so it's possible to skip heavy test scenarios which are not stable when
running against binaries built with ASAN/TSAN support.

I didn't add any tests, but this functionality is going to be used
in a follow-up patch.

Change-Id: I08c978ecda7321038b97c5c885914c02122606c1
Reviewed-on: http://gerrit.cloudera.org:8080/17487
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>