[consensus] a small optimisation on log prefix
This patch introduces a small optimisation into the implementation of
the RaftConsensus::LogPrefixThreadSafe() method: since the prefix
it produces doesn't change, it's possible to construct the string once
in the constructor and then return a constant reference.

I also updated PendingRound class to store a reference to the prefix as
well since instances of the PendingRound class have shorter live span
than RaftConsensus objects which create PendingRound objects.

Change-Id: I6a5082fe07449ef24827696638ddd8e925a08264
Reviewed-on: http://gerrit.cloudera.org:8080/17489
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>