[txn_manager] allow a super user to call txn-related RPCs
Since we are planning to use transactional operations in kudu CLI tools
(e.g., in 'kudu perf loadgen'), it makes sense to allow txn-related RPCs
to be invoked by a super user as well since some kudu CLI tools are often
run under such credentials.  This patch changes the RPC-level coarse authz
settings for all methods in the TxnManagerService accordingly.

I also had a version of this patch where a service user was able to call
txn API as well, but it turned out that TabletServerService API isn't
accessible to service users, e.g. TabletServerService::Write() is
accessible only to regular and super users.  With that, I don't quite see
the necessity of allowing a service user to call txn API.

Change-Id: I21e4b29634fd01f0ced80b54373b1ce156e274ad
Reviewed-on: http://gerrit.cloudera.org:8080/17510
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>