[test] minor clean up on TestUnsafeChangeConfigLeaderWithPendingConfig
When looking at one of recent pre-commit test failure [1][2], I found
that AdminCliTest.TestUnsafeChangeConfigLeaderWithPendingConfig failed
but it was hard to tell what was the actual status code returned:

  src/kudu/tools/kudu-admin-test.cc:881: Failure
  Value of: s.IsTimedOut()
    Actual: false
  Expected: true

I reran the scenario multiple times with TSAN-enabled binaries, but
was not able to reproduce the test failure yet.  Anyways, I added more
diagnostics about the actual status code returned and cleaned up the
code of the test scenario a bit.  Hopefully, next time it fails it
will be clearer what's going on.

[1] http://jenkins.kudu.apache.org/job/kudu-gerrit/23918/BUILD_TYPE=TSAN
[2] http://dist-test.cloudera.org/job?job_id=jenkins-slave.1621893605.3746155

Change-Id: Id2bf354764f3ddb371de19910c5f879c1a06c78b
Reviewed-on: http://gerrit.cloudera.org:8080/17504
Tested-by: Kudu Jenkins
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>